support sub-second timeouts SERVER-1515

+++ util/message.h
@@ -97,7 +97,7 @@ namespace mongo {
         // in some cases the timeout will actually be 2x this value - eg we do a partial send,
         // then the timeout fires, then we try to send again, then the timeout fires again with
         // no data sent, then we detect that the other side is down
-        MessagingPort(int timeout = 0, int logLevel = 0 );
+        MessagingPort(double timeout = 0, int logLevel = 0 );
 
         virtual ~MessagingPort();
 
@@ -132,7 +132,7 @@ namespace mongo {
         PiggyBackData * piggyBackData;
     public:
         SockAddr farEnd;
-        int _timeout;
+        double _timeout;
         int _logLevel; // passed to log() when logging errors
 
         static void closeAllSockets(unsigned tagMask = 0xffffffff);
